Put JMH substitutions into one class
